examples/autonomous_agents/devops_telegram_bot/requirements.txt
upsonic
python-dotenv
fastapi
uvicorn
sqlalchemy
aiosqlite
anthropic
